pyproject.toml
[project]
name = "cognitive-os"
version = "0.1.0"
description = "C-OS: Persistent temporal cognitive operating system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "numpy>=1.26.0",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
neo4j = ["neo4j>=5.24.0"]
faiss = ["faiss-cpu>=1.8.0"]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["cos*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
